Extract xlsx hidden data detection
Moving into its own module so we can reuse this code for scanning `xls`
files and maybe even raw emails too.
